Fix and test.
