+++ allennlp/tests/training/trainer_test.py
@@ -11,7 +11,7 @@ import torch
 import pytest
 from allennlp.common.checks import ConfigurationError
 
-from allennlp.common.testing import AllenNlpTestCase
+from allennlp.common.testing import AllenNlpTestCase, ModelTestCase
 from allennlp.training import Trainer
 from allennlp.training.trainer_base import TrainerBase
 from allennlp.training.learning_rate_schedulers import LearningRateScheduler
@@ -494,3 +494,13 @@ class TestSparseClipGrad(AllenNlpTestCase):
         # Final norm should be 1.5
         grad = embedding.weight.grad.coalesce()  # pylint: disable=no-member
         self.assertAlmostEqual(grad._values().norm(2.0).item(), 1.5, places=5) # pylint: disable=protected-access
+
+class TestLanguageModelWithMultiprocessDatasetReader(ModelTestCase):
+    def setUp(self):
+        super().setUp()
+        self.set_up_model(self.FIXTURES_ROOT / 'language_model' / 'experiment_multiprocessing_reader.jsonnet',
+                          # Note the glob on the end of this path.
+                          self.FIXTURES_ROOT / 'language_model' / 'sentences*')
+
+    def test_unidirectional_language_model_can_train_save_and_load(self):
+        self.ensure_model_can_train_save_and_load(self.param_file)
